Adding Careem Licence and extract logic
* Adding careem licence to all files where it was missing
* Extracted generation logic form the plugin itself
